Add an example screenshotter using export-dmabuf
This takes a screenshot for each display and saves to `.png` images.

I'm not entirely sure the dmabuf/OpenGL logic here handles formats. I
assume OpenGL can be trusted to convert any DRM format to 8 bit per
channel uncompressed RGBA? Which could be lossy depending on the format?

It could be good to have examples testing other protocols. And
something testing dmabuf exporting for workspaces and toplevels. (It
could similar iterate over workspaces and toplevels and export an image
for each.)