client-toolkit/toplevel_info: Ignore `done` before cosmic events
The compositor sends a `done` event for the
`ext_foreign_toplevel_handle_v1` before we call `get_cosmic_toplevel`.
Ignore that and wait for the `done` that occurs after we get cosmic
events.

`zcosmic_toplevel_handle_v1::state` seems to be a mandatory event, so it
is valid if awkward to use for the purpose. I don't know about a cleaner
solution. We could just ignore the first `done`, but technically it
could race if the compostior sends a state change before it has
processed the `get_cosmic_toplevel`. Maybe a roundtrip/sync could work.