client-toolkit/screencopy: Improved API with new types
This adds a `CaptureSource` enum, similar to what cosmic-workspaces and
xdg-desktop-portal-cosmic, to construct session without directly
accessing manager types. It also adds a ref-counted `CaptureSession`
that destroys the session when the last reference is dropped.

This is a step towards abstracting support for the upstream
`ext-image-copy-capture-v1`. Otherwise it just reduces the code required
by callers a little bit, but more high level helpers should probably
be added as well.