client-toolkit/toplevel-info: Work with only ext foreign toplevel
With this, `toplevel-monitor` works on Sway, but doesn't have the
information exposed through the Cosmic protocol, naturally.

The API is changed to use `ExtForeignToplevelHandleV1`. For convenience,
`ToplevelInfo` contains the foreign handle, and the optional cosmic
handle.

I assume `identifier` is meant to be mandatory, so that is no longer an
`Option` here. Though there's no check that the server sends this (is
server-side code, this would send a protocol error; I guess cctk could
panic, or log an error.)