client-toolkit/screencopy: Map foreign toplevel to cosmic toplevel
This removes `CaptureSource::CosmicToplevel`, and merges it with
`CosmicSource::Toplevel`. We already require a version of
cosmic-toplevel based on foreign toplevel, so this mapping should always
be possible as long as the cosmic toplevel handle was created by the
same version of cctk (which shouldn't be an issue).

This means callers don't have to map cosmic toplevels to foreign
toplevels, or do anything special to support both `CosmicToplevel` and
`Toplevel` capture sources.

To be able to access the mapping without access to state, including from
a different thread, this adds user data to `ExtForeignToplevelHandle`.
This is a bit annoying to handle with `event_created_child`, but a
`OnceLock` seems like as good a way as any to do this.