Add capture sources for `ext_workspace_handle_v1`
This adds a global for creating a `zcosmic_image_source_v1`, *and* one
for creating an `ext_image_capture_source_v1`, from an
`ext_workspace_handle_v1`. This is a prerequisite for migrating away
from the current Cosmic workspace protocol.

In cctk, this adds a `CaptureSource::Workspace`, that then works with
both ext and cosmic screen capture, if the necessary global exists in
either case.

Once cosmic-comp is updated to support `ext-image-copy-capture-v1`, the
legacy source global can be removed. But it's easy enough to add for now
for migration. Then we can also propose a workspace capture source
upstream.