Add ext workspace requests/events to toplevel info/management
This is needed to migrate from `cosmic-workspace-unstable-v1` to the new
upstream protocol.

Duplicating methods/events with `ext_` in the name is a bit annoying,
but this is ultimately pretty easy and clean to support. If the
compositor drops `cosmic-workspace-unstable-v1` support, these methods
simply become inert, since no objects of that handle type will exist.

Particularly with that, there's no particular need for the old version
to not be supported on the new protocol version. So there's no such
restriction here.

For now, the cctk changes are non-breaking, and just add the ext
workspaces separately.