client-toolkit/toplevel_info: Remove cosmic workspace v1 handles
The workspace helper now requires the v2 protocol, and ext-workspaces,
so this is no longer relevant.

I thought about bumping the minimum version for this an toplevel
management, but there's not really any need to restrict that for clients
that don't bind the workspaces protocol.

Regardless, now everything is using ext workspace handles.