client-toolkit/workspace: Don't send done until cosmic info received
And don't commit the double-buffered state.

This is similar to the `toplevel_info` handler, though it applies to
all workspaces and groups.

We could send groups, and any workspace that does have cosmic info, but
it seems better to wait a roundtrip than to have the client handle
changes to some but not other workspaces.