client-toolkit/workspace: Add workspace/group `_info` methods
Equivalent to the `ToplevelInfoState::info` method for toplevels. So now
the workspace and toplevel helpers have similar APIs.

This is quite helpful now that workspace groups only contain a set of
workspace handles, not info, and need to be looked up.

(From a performance standpoint, just a linear search through an array is
likely as fast as anything with the number of workspaces that are likely
to exist, so this implementation should be fine.)