output-management: Allow config of xrandr primary output
Having a boolean value, that is only valid if set for at most a single
output in the protocol is weird, as the protocol doesn't specify, that
a full state has to be sent. It treats each head as being independant
from the next.

As such cosmic-randr would have to sent `set_xwayland_primary` for
all heads even unchanged ones to update the state without an
`already_set` error (or keep track of the previous head), if we
added the request to `zcosmic_output_configuration_head_v1`.

Instead I opted to model this with a separate request on the
`zcosmic_output_manager_v1` as this isn't part of the drm-state and
thus doesn't need to be set atomically anyway.